Fix and finish function to find line intersections
As it is now, this code makes it possible to implement an algorithm to find nearby points by making rays and checking if they intersect with every existing wall, needing m*n operations.

That's fine, albeit I'd like to try the digital differential analyzer algorithm as well, so that'll be the next thing to implement.